sentence_vectorizer_api/requirements.txt
gensim
pyspellchecker